agents/requirements.txt
google-adk
pydantic==2.12.5
